+++ publish/CHANGELOG.md
@@ -1,5 +1,7 @@
 - v1.18
   - Adds spawn specific triggerRules to the field `spawn`.
+  - Adds new fields `addItems` and `removeItems` to more easily add or remove items from containers.
+  - Adds support for filtering with the data type items.
 
 - v1.17
   - Adds new filters `paint`, `minPaint` and `maxPaint`.